+++ packages/sycamore-core/src/lib.rs
@@ -9,6 +9,7 @@
 
 pub mod component;
 pub mod generic_node;
+#[cfg(feature = "hydrate")]
 pub mod hydrate;
 pub mod noderef;
 pub mod render;
