+++ packages/sycamore-web/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "sycamore-web"
 categories = ["gui", "wasm", "web-programming"]
-description = "Web utilities for the Sycamore UI framework"
+description = "Web renderer for the Sycamore UI framework"
 edition = "2021"
 homepage = "https://github.com/sycamore-rs/sycamore"
 keywords = ["wasm", "gui", "reactive", "web"]
@@ -11,9 +11,33 @@ repository = "https://github.com/sycamore-rs/sycamore"
 version = "0.8.0-beta.5"
 
 [dependencies]
+html-escape = "0.2.11"
+indexmap = "1.8.1"
+js-sys = "0.3.57"
+once_cell = "1.10.0"
 sycamore-core = { path = "../sycamore-core", version = "0.8.0-beta.5" }
 sycamore-reactive = { path = "../sycamore-reactive", version = "0.8.0-beta.5" }
 wasm-bindgen = "0.2.80"
+web-sys = { version = "0.3.57", features = [
+    "console",
+    "Comment",
+    "Document",
+    "DocumentFragment",
+    "DomTokenList",
+    "Element",
+    "Event",
+    "HtmlElement",
+    "HtmlInputElement",
+    "HtmlCollection",
+    "Node",
+    "NodeList",
+    "Text",
+    "Window",
+] }
+
+[dev-dependencies]
+sycamore = { path = "../sycamore" }
 
 [features]
-ssr = []
+hydrate = ["sycamore-core/hydrate"]
+ssr = ["hydrate"]
